fix: CI 不能把 tests 当包导入夹具
夹具抽到 xlsx_fixtures.py，pytest pythonpath 带上 tests。
