+++ pyproject.toml
@@ -32,7 +32,7 @@ packages = ["src/docparse"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
-pythonpath = ["src"]
+pythonpath = ["src", "tests"]
 
 [tool.ruff]
 line-length = 100
